pyproject.toml
[tool.poetry]
name = "pizzeria"
version = "0.1.0"
description = ""
authors = ["Sergio Gómez <decano@gmail.com>"]
license = "EUPL-1.2"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.83.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
loguru = "^0.6.0"
tortoise-orm = {extras = ["asyncpg", "asyncmy", "asyncodbc"], version = "^0.19.2"}
aerich = "^0.6.3"
ulid-py = "^1.1.0"
strawberry-graphql = "^0.129.0"
result = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-sugar = "^0.9.5"
pytest-xdist = "^2.5.0"
pytest-asyncio = "^0.19.0"
black = {version = "^22.6.0", allow-prereleases = true}
mypy = "^0.971"
debugpy = "^1.6.3"
pytest-bdd = "^6.0.1"
requests = "^2.28.1"
isort = "^5.10.1"
pytest-env = "^0.6.2"
nest-asyncio = "^1.5.5"
types-requests = "^2.28.9"
Faker = "^14.1.0"
PyHamcrest = "^2.0.4"
pre-commit = "^2.20.0"

[tool.poetry.scripts]
serve = "pizzeria.kernel:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
flake8-docstrings = "^1.6.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "pizzeria.system.database.ENV"
location = "./migrations"
src_folder = "./."

[tool.pytest.ini_options]
bdd_features_base_dir = "features/"
asyncio_mode = "auto"

[tool.isort]
profile = "black"
line_length = 110

[tool.black]
line-length = 110
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
follow_imports = "skip"
ignore_missing_imports = true
no_implicit_optional = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_none_errors = true
show_traceback = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
plugins = [
  "strawberry.ext.mypy_plugin",
]
files = [
  "src",
]
